prediction/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
catboost>=1.2
xgboost>=2.0.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
influxdb3-python>=0.1.0
joblib>=1.3.0
scikit-learn>=1.3.0
httpx>=0.26.0
